UI improvements
A pretty large changeset. Introduces the concept of "profiles",
which allow the user to store their settings in a file for use
on multiple computers. This also makes it easier for multiple
users to use the program on the same system.

As a way of making these changes possible, we also move to a
traditional menu bar interface. Creating / opening / editing
both a profile and a data file proved too complicated for the
toolbar approach to work.

Contains a number of minor improvements and cleanups that
happened in the process of working on the above.