improve mpl font rendering
It seems MPL uses the auto-hinter unless you set contrary options,
which is a very strange choice. Unfortunately it doesn't provide
an option to use the platform's default hinting settings, so instead
we choose to use a default that looks good on most Linux systems.

Also, override the MPL default choice of DejaVu Sans.